🩹 refactor : 충돌 해결 및 상품 머지
