+++ src/main/java/com/sparta/itsmine/domain/product/entity/Product.java
@@ -1,5 +1,7 @@
 package com.sparta.itsmine.domain.product.entity;
 
+import static com.sparta.itsmine.domain.product.utils.ProductStatus.SAVED;
+
 import com.sparta.itsmine.domain.category.entity.Category;
 import com.sparta.itsmine.domain.product.utils.ProductStatus;
 import com.sparta.itsmine.domain.user.entity.User;
@@ -25,6 +27,7 @@ import lombok.NoArgsConstructor;
 @Table(name = "product")
 @NoArgsConstructor
 public class Product extends TimeStamp {
+
     /**
      * 컬럼 - 연관관계 컬럼을 제외한 컬럼을 정의합니다.
      */
@@ -81,7 +84,7 @@ public class Product extends TimeStamp {
         this.category = category;
 
         // set up initialized values
-        this.status = ProductStatus.BID;
+        this.status = SAVED;
         this.startDate = LocalDateTime.now();
         this.like = false;
     }
