--- src/main/java/com/sparta/itsmine/domain/product/entity/ProductRepository.java
@@ -1,7 +1,0 @@
-package com.sparta.itsmine.domain.product.entity;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ProductRepository extends JpaRepository<Product,Long> {
-
-}
